// src/app/Register/Model/user.ts
export interface User {
  userName: string;
  firstName: string;
  lastName: string;
  phone: string;
  email: string;
  dob: string;
  courseId: string;
  fatherOccupation: string;
  fatherName: string;
  fPhone: string;
  fatherOrg: string;
  fAnnualIncome: string;
  motherName: string;
  motherOccupation: string;
  motherOrg: string;
  mAnnualIncome: string;
}